add qq to fill_missing

+++ workflow/scripts/common/utils.py
@@ -473,9 +473,9 @@ def get_fst_faser(ts, popA, popB):
 	return(fst)
 
 
-def get_ancestry_dosage(arr, n_anc):
+def get_ancestry_dosage(arr, n_anc, dtype=np.float32):
 	"""Compute ancestry dosage from probablistic haploid ancestry calls."""
-	anc_dosage = np.zeros((arr.shape[0], int(arr.shape[1] / 2)), dtype=np.half)
+	anc_dosage = np.zeros((arr.shape[0], int(arr.shape[1] / 2)), dtype=dtype)
 	# anc_dosage = np.zeros((arr.shape[0], int(arr.shape[1] / 2)))
 	if n_anc == 2:
 		a0 = arr[:, 0::2]  # should be views
